added sketched app storyboard for world planning
This is a sketch of a speculated VR social platform for civic / sustainability planning on local and global scales. Harnessing creativity, this platform functions as a facsimile of the world that can be modified by users. I'll be adding more sketches that illustrate the idea for the planning app - this is not something we will be making within this project - but just to give a scope for the possibilities of how planning in VR could develop.